[refactor] step2 : 컴포넌트 분리
